- Moved Background and BackgroundData into a subdir - Added dummy gfx, when a requested resource couldn't be found - Pseudo Sound goes to stdout - Fixed bug in the < operator of ResDescriptor - Added a greater catch radius for the blocker - Fixed the miner goes through solid bug

+++ src/backgrounds/ThunderstormBackground.cc
@@ -1,4 +1,4 @@
-//  $Id: ThunderstormBackground.cc,v 1.1 2000/08/28 00:23:50 grumbel Exp $
+//  $Id: ThunderstormBackground.cc,v 1.2 2000/09/07 09:45:39 grumbel Exp $
 //
 //  Pingus - A free Lemmings clone
 //  Copyright (C) 2000 Ingo Ruhnke <grumbel@gmx.de>
@@ -43,7 +43,7 @@ void
 ThunderstormBackground::let_move(void)
 {
   x_pos += 1;
-  if (x_pos >= clouds_sur->get_width())
+  if (x_pos >= (int) clouds_sur->get_width())
     x_pos = 0;
 }
